refactor fork/exec logic for the command runners
